UserServiceインターフェース作成、既存のUserServiceの実装クラス化。未動作確認
