+++ src/main/java/com/example/app/service/UserService.java
@@ -1,44 +1,12 @@
 package com.example.app.service;
 
-import org.springframework.stereotype.Service;
-
 import com.example.app.domain.User;
-import com.example.app.mapper.PasswordMapper;
-import com.example.app.mapper.UsersMapper;
 
-import lombok.RequiredArgsConstructor;
-
-@Service
-@RequiredArgsConstructor
-public class UserService {
-    
-    private final UsersMapper usersMapper;
-    private final PasswordMapper passwordMapper;
+public interface UserService {
 
     //ユーザー登録
-    public void registerUser(User user){
-        //ユーザーテーブルに追加
-        usersMapper.addUser(user);
-
-        //パスワードテーブルに追加
-        user.setId(user.getId()); //addUserでセットしたID
-        passwordMapper.addPassword(user);
-    }
-
+    public void registerUser(User user);
+    
     //ログイン処理
-    public boolean canUserLogin(User user){
-        //ユーザーネームからID、パスワードを取得
-        Integer userId = usersMapper.selectByName(user.getUserName());
-        if(userId == null){
-            return false;
-        }
-
-        String password = passwordMapper.selectById(userId);
-
-        //DBから取得したパスワードと入力フォームのパスワードを比較
-        if(password.equals(user.getUserPassword())){
-            return true;
-        }
-        return false;
-    }
+    public boolean canUserLogin(User user);
 }
